requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
Pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0 